Add missing files, update min
