+++ holotapes/Book-It/ChangeLog
@@ -1,0 +1,3 @@
+1.0.0
+https://github.com/CodyTolene/pip-boy-3000-holotapes/pull/74
+- Initial release
