+++ holotapes/Book-It/README.md
@@ -11,17 +11,21 @@
 - [Pip-Boy](https://www.Pip-Boy.com)
 - [Pip-Boy Apps](https://github.com/CodyTolene/pip-boy-apps)
 
-###Description
+### Description
 
-- A .txt file reader for your Pip-Boy. Supports subfolders to help organize, and brightness control with the left wheel while reading. Auto-bookmarking per page change, creates a .sav file in same folder as the .txt file you read.  Loads flavor text from the subs.js file located in assets, goes in same folder as app.js when installed.
+A .txt file reader for your Pip-Boy. Supports subfolders to help organize, and
+brightness control with the left wheel while reading. Auto-bookmarking per page
+change, creates a .sav file in same folder as the .txt file you read. Loads
+flavor text from the subs.js file located in assets, goes in same folder as
+app.js when installed.
 
 ### Controls
 
-| Input                | Action                                                                |
-| -------------------- | --------------------------------------------------------------------- |
-| Left wheel (rotate)  | Scrolls .txt files in browser and change brightness while reading     |
-| Left wheel (press)   | Select .txt file in browser, returns to browser if in reader          |
-| Right wheel (rotate) | Scrolls .txt files in browser, changes page while reading             |
+| Input                | Action                                                            |
+| -------------------- | ----------------------------------------------------------------- |
+| Left wheel (rotate)  | Scrolls .txt files in browser and change brightness while reading |
+| Left wheel (press)   | Select .txt file in browser, returns to browser if in reader      |
+| Right wheel (rotate) | Scrolls .txt files in browser, changes page while reading         |
 
 ### Credits
