+++ holotapes/Book-It/assets/SUBS_.JS
@@ -1,0 +1,102 @@
+[
+  'Knowledge is power!',
+  'Got a library card?',
+  'Now get to reading!',
+  'Intelligence +1',
+  'Fiction or non-fiction?',
+  'Now with 57% less paper!',
+  "You wouldn't steal a Holotape.",
+  'A product of Vault 919',
+  'You load 16 tons, and what do you get?',
+  'Read it again, my Johnny.',
+  'I Have No Book and I Must Read',
+  'U, U, D, D, L, R, L, R, B, A, S',
+  '[ Hidden ]',
+  '[ Caution ]',
+  '[ Danger ]',
+  'You take a sip from your trusty Vault 13 canteen.',
+  'War. War never changes.',
+  'You cannot fast travel when enemies are nearby.',
+  'You cannot fast travel when overencumbered.',
+  'I came, I saw, I looted.',
+  "Ain't that a kick in the head!",
+  'Moo, I say.',
+  'Please assume the position.',
+  "Welcome to Camp Navarro. So, you're the new replacement.",
+  'I have a theoretical degree in physics!',
+  'Drink. Some. Water.',
+  "Don't mind me, just enjoying the view.",
+  'Give me that chip.',
+  'LIGAF',
+  'Believe in the you that I believe in!',
+  'Then who was phone?',
+  "I am serious, and don't call me Shirley.",
+  'Now, walk this way.',
+  'Always look on the bright side of life!',
+  'He never wants a second cup of coffee at home...',
+  'Be prepared.',
+  'We apologize for the inconvenience.',
+  "DON'T PANIC",
+  'Everything is connected.',
+  "I'm not a fan of reality, it's a bit overrated.",
+  "Pining for the f'jords.",
+  'Oh no, not again...',
+  "I know what we're gonna do today!",
+  "But wait, there's more!",
+  'Pip-Boy.com',
+  'robco-industries.org',
+  'thewandcompany.com',
+  'GARY!',
+  'Another settlement needs your help.',
+  'Your Pip-Boy has detected a signal.',
+  'I had something for this...',
+  'What is your spaghetti policy here?',
+  'Rock! Flag! Eagle!',
+  'Wild card baby! YEEEHAAAW!!!',
+  "I'm well versed in bird law.",
+  'So anyways, I started reading.',
+  'Please insert disk 2.',
+  'Loading from cassette... Please wait.',
+  "Press any key to continue... or don't.",
+  'A winner is you!',
+  'All your base are belong to us.',
+  'Do a barrel roll!',
+  'The cake is a lie.',
+  "It's a trap!",
+  'Have you tried turning it off and on again?',
+  'Initiating self-destruct sequence in 3,2,1...',
+  "Don't feed the Deathclaws.",
+  'Error 404: Text not found',
+  'Task failed successfully.',
+  'I see you shiver with antici.....',
+  '...pation!',
+  "It's just a jump to the left!",
+  "It's super effective!",
+  'The path is blocked by a small shrubbery.',
+  "Don't dream it, be it.",
+  "It's not easy having a good time.",
+  "I'd like, if I may, to take you on a strange journey.",
+  "You're not looking at a king. You're looking at an ace.",
+  "I've just come to tell you how fabulous I am.",
+  'The sun in the sky is a spotlight just for us.',
+  'Never knows best',
+  "Nothing can happen 'til you swing the bat.",
+  "Sometimes you're the cat, sometimes the mouse.",
+  'Congratulations!',
+  'Kick reason to the curb and do the impossible!',
+  "A true man never dies, even when he's killed!",
+  'Evolve beyond the person you were one minute before.',
+  'Aim for the Top!',
+  "Let's see you grit those teeth!",
+  'Time waits for no one.',
+  'Activate the Ide!',
+  'Because of the implication.',
+  'Can I offer you a nice egg in this trying time?',
+  'You gotta pay the troll toll.',
+  'Derivative!',
+  "You haven't thought of the smell!",
+  'Science is a liar sometimes.',
+  'Mostly harmless.',
+  "I'd far rather be happy than right any day.",
+  'Reality is frequently inaccurate.',
+];
