+++ holotapes/registry.json
@@ -166,6 +166,24 @@
     ]
   },
   {
+    "id": "BOOKIT",
+    "name": "Book-It",
+    "author": "@Theeohn",
+    "version": "1.0.0",
+    "description": "A .txt file reader for your Pip-Boy. Auto-bookmarking per page change, supports subfolders to help organize, and brightness control with the left wheel while reading.",
+    "icon": "Book-It/assets/book.png",
+    "previews": [],
+    "type": "app",
+    "readme": "Book-It/README.md",
+    "storage": [
+      {
+        "name": "HOLO/BOOK_IT/app.js",
+        "url": "Book-It/app.min.js"
+      }
+    ],
+    "storageOptional": []
+  },
+  {
     "id": "checkers",
     "name": "Checkers",
     "author": "@trekker87 @CodyTolene",
